Refactored file layout, splitting export.go into several smaller files according to their function export.go -> controller links.go -> logic related to generating links in status pvc-source -> if source is a PVC vm-source -> if souce is a VM vmsnapshot-source -> if source is a VMSnapshot
Also addressed review comments.
